[#150] 여행 일정 작성 api 연결

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/BookmarkedPlace.kt
@@ -1,6 +1,6 @@
 package kr.tekit.lion.daongil.domain.model
 
 data class BookmarkedPlace(
-    val bookmarkedPlaceId : Int,
+    val bookmarkedPlaceId : Long,
     val bookmarkedPlaceName : String
 )
